Revert DonateFiatThanks
